Set up the use of clang format and clang tidy
