+++ 2023/BUILD
@@ -2,28 +2,31 @@ load("@rules_cc//cc:defs.bzl", "cc_binary", "cc_test", "cc_library")
 
 cc_library(
     name = "common",
-    hdrs = ["src/common.h"],
+    hdrs = ["src/common.hpp"],
     includes = ["src"],
 )
 
 cc_library(
     name = "day1",
-    srcs = ["src/day1/day1.cc"],
-    hdrs = ["src/day1/day1.h"],
+    srcs = ["src/day1/day1.cpp"],
+    hdrs = ["src/day1/day1.hpp"],
     deps = [":common"],
 )
 
 cc_binary(
     name = "aoc2023",
-    srcs = ["src/main.cc"],
+    srcs = ["src/main.cpp"],
     deps = [":day1"],
 )
 
 cc_test(
     name = "aoc2023_test",
-    srcs = ["test/test.cc"],
+    srcs = ["test/test.cpp"],
     deps = [
         "@catch2//:catch2_main",
         ":day1",
+    ],
+    data = [
+        "src/day1/day1.testinput",
     ],
 )
