--- 2023/src/day1/day1.cc
@@ -1,4 +1,0 @@
-#include "day1.h"
-
-int Day1Solution::part1(std::istream &in) const { return 0; }
-int Day1Solution::part2(std::istream &in) const { return 0; }
