--- 2023/src/day1/day1.h
@@ -1,3 +1,0 @@
-#include "common.h"
-
-SOLUTION(1, 24000, 45000);
